Move process_onion_failure into onion_utils
